mcp/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gundelsby-mcp"
version = "0.1.0"
description = "Model Context Protocol server exposing gundelsby.com content"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "python-frontmatter>=1.0.1",
    "PyYAML>=6.0.0"
]
